Implement cli loadwallet command
